# backendPI4/api/openapi.yaml
openapi: 3.0.0
info:
  title: API para a aplicação ViriatosScouting
  description:
    "Este documento apresenta toda a documentação de uma API que permite\
    \ gerir um sistema que faz gestão de uma equipa de scouting. A API inclui funcionalidades\
    \ que permitem o utilizador gerir jogadores, clubes, escalões, equipas e utilizadores.\
    \ A mesma tambem apresenta um local para criar tarefas e relatórios."
  version: "1.0"
servers:
  - url: https://localhost:8080
    description: Servidor local utilizado para testar a API
  - url: https://backendpi4-bbzf.onrender.com
    description: Servidor online utilizado para testar a API
...
paths:
...
  /tarefas:
    get:
      tags:
        - Tarefas
      summary: Lista todos os tarefas
      description:
        "Retorna uma lista de todas as tarefas, mas somente utilizadores\
        \ autenticados (administradores e olheiros) podem visualizar esta informaç\
        ão. Para obter as tarefas, o utilizador deve estar autenticado com um token\
        \ válido."
      operationId: listaTarefas
      parameters:
        - name: limit
          in: query
          description: Número máximo de tarefas a serem retornadas por página.
          required: false
          style: form
          explode: true
          schema:
            maximum: 999999
            minimum: 1
            type: integer
            format: int64
            example: 10
            default: 10
        - name: offset
          in: query
          description: Número de tarefas a ignorar para paginação.
          required: false
          style: form
          explode: true
          schema:
            maximum: 300
            minimum: 0
            type: integer
            format: int64
            example: 1
            default: 0
        - name: idUtilizador
          in: query
          description: Filtra as tarefas pelo identificador único do utilizador.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int64
            example: 3
        - name: idAtleta
          in: query
          description: Filtra as tarefas pelo identificador único do atleta.
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int64
            example: 3
      responses:
        "200":
          description: Lista de tarefas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarTarefa200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma tarefa encontrada para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorTarefa"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
    post:
      tags:
        - Tarefas
      summary: Cria uma nova tarefa
      description:
        Permite criar uma nova tarefa. Apenas administradores podem criar
        tarefas. O utilizador deve fornecer os dados da tarefa no body.
      operationId: criarTarefa
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TarefaCriar"
        required: true
      responses:
        "201":
          description: Tarefa criada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarefaCriar"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
  /tarefas/{id}:
    get:
      tags:
        - Tarefas
      summary: Lista uma tarefa
      description: "Retorna uma lista de uma tarefa, mas somente utilizadores\
        \ autenticados (administradores e olheiros) podem visualizar esta informaç\
        ão. Para obter as tarefas, o utilizador deve estar autenticado com um token\
        \ válido."
      operationId: listaUmaTarefa
      parameters:
        - name: id
          in: path
          description: ID da tarefa a ser atualizada
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      responses:
        "200":
          description: Lista de tarefas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarUmaTarefa200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma tarefa encontrada para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorTarefa"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
    put:
      tags:
        - Tarefas
      summary: Edita um tarefa
      description:
        Permite editar um tarefa existente. O utilizador pode editar tarefas
        se o mesmo for administrador. É necessário um token válido para autenticação.
      operationId: editarTarefa
      parameters:
        - name: id
          in: path
          description: ID da tarefa a ser atualizada
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TarefaCriar"
        required: true
      responses:
        "200":
          description: Tarefa atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarefaCriar"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
    delete:
      tags:
        - Tarefas
      summary: Apaga um tarefa
      description:
        Permite apagar um tarefa existente. Apenas o administrador pode
        apagá-la. O utilizador deve fornecer o ID do tarefa a ser apagada.
      operationId: apagarTarefa
      parameters:
        - name: id
          in: path
          description: ID do tarefa a ser apagada
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      responses:
        "204":
          description: Tarefa apagado com sucesso
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhum tarefa encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorTarefa"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
  /tarefas/proxima:
    get:
      tags:
        - Tarefas
      summary: Lista proxima tarefa de um utilizador
      description:
        "Retorna a proxima tarefa de um utilizador, mas somente utilizadores\
        \ autenticados (olheiros) podem visualizar esta informaç\
        ão. Para obter as tarefas, o utilizador deve estar autenticado com um token\
        \ válido."
      operationId: listaProximaTarefa
      responses:
        "200":
          description: Lista de tarefas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarUmaTarefa200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma tarefa encontrada para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorTarefa"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
  /tarefas/proximas:
    get:
      tags:
        - Tarefas
      summary: Lista proximas tarefas de um utilizador
      description:
        "Retorna a proxima tarefa de um utilizador, mas somente utilizadores\
        \ autenticados (olheiros) podem visualizar esta informaç\
        ão. Para obter as tarefas, o utilizador deve estar autenticado com um token\
        \ válido."
      operationId: listaProximasTarefas
      responses:
        "200":
          description: Lista de tarefas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarUmaTarefa200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma tarefa encontrada para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorTarefa"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Tarefas
...
components:
  schemas:
...
    AtletaTarefa:
      required:
        - clube_id
        - id
        - nome
      type: object
      properties:
        id:
          maximum: 999999
          minimum: 1
          type: integer
          description: Identificador único do atleta
          format: int64
          example: 1
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome completo do atleta
          example: Cristiano Ronaldo
        clube_id:
          $ref: "#/components/schemas/Clube"
      additionalProperties: false
...
    Clube:
      required:
        - id
        - nome
      type: object
      properties:
        id:
          maximum: 999999
          minimum: 1
          type: integer
          description: Identificador único do clube
          format: int64
          example: 1
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome do clube
          example: Rocas do Vouga
        icon:
          maxLength: 512
          pattern:
            "^(https?|ftp):\\/\\/[\\w\\-]+(\\.[\\w\\-]+)+(\\/[\\w\\-\\.~:@\\\
            /?#\\[\\]@!\\$&'\\(\\)\\*\\+,;=%]*)?$"
          type: string
          description: URL ou caminho para o ícone/imagem do atleta
          example: https://example.com/images/sporting_logo.png
      additionalProperties: false
      example:
        icon: https://example.com/images/sporting_logo.png
        nome: Rocas do Vouga
        id: 1
...
    Tarefa:
      required:
        - atleta_id
        - idClube
        - data
        - id
        - utilizador_id
      type: object
      properties:
        id:
          type: integer
          description: Identificador único da tarefa
          format: int64
          example: 1
        data:
          type: string
          description: Data da tarefa (formato YYYY-MM-DD)
          format: date
          example: 2024-12-29
        horaInicio:
          type: string
          description: Hora de início da tarefa (formato YYYY-MM-DDTHH:MM:SS)
          format: date-time
          example: 2024-11-20T09:00:00Z
        horaFim:
          type: string
          description: Hora de fim da tarefa (formato YYYY-MM-DDTHH:MM:SS)
          format: date-time
          example: 2024-11-20T11:00:00Z
        jogaCasa:
          type: boolean
          description: Indica se a equipa joga em casa (true) ou fora (false)
          example: true
        idClube:
          $ref: "#/components/schemas/Clube"
        idUtilizador:
          $ref: "#/components/schemas/UtilizadorTarefa"
        idAtleta:
          $ref: "#/components/schemas/AtletaTarefa"
      additionalProperties: false
    TarefaCriar:
      required:
        - idAtleta
        - idClube
        - data
        - idUtilizador
      type: object
      properties:
        data:
          type: string
          description: Data da tarefa (formato YYYY-MM-DD)
          format: date
        horaInicio:
          type: string
          description: Hora de início da tarefa (formato YYYY-MM-DDTHH:MM:SS)
          format: date-time
          example: 2024-11-20T09:00:00Z
        horaFim:
          type: string
          description: Hora de fim da tarefa (formato YYYY-MM-DDTHH:MM:SS)
          format: date-time
          example: 2024-11-20T11:00:00Z
        idClube:
          type: integer
          description: Identificador único do clube relacionado com a tarefa
          format: int64
          example: 1
        idUtilizador:
          type: integer
          description: Identificador único do utilizador relacionado com a tarefa
          format: int64
          example: 2
        idAtleta:
          type: integer
          description: Identificador único do atleta relacionado com a tarefa
          format: int64
          example: 3
        jogaCasa:
          type: boolean
          description: Indica se a equipa joga em casa (true) ou fora (false)
          example: true
      additionalProperties: false
...
    UtilizadorTarefa:
      required:
        - nome
      type: object
      properties:
        id:
          type: integer
          description: Identificador único do utilizador
          format: int64
          example: 1
        nome:
          maxLength: 512
          type: string
          description: Nome do utilizador
          example: João Silva
      additionalProperties: false
...
    UnexpectedError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9À-ÿ ]+$"
          type: string
          example: Ocorreu um erro inesperado no servidor.
      additionalProperties: false
    BadRequestError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example: Token de autenticação inválido ou ausente.
      additionalProperties: false
    UnauthorizedError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example: Campos obrigatórios ausentes ou formato inválido.
      additionalProperties: false
...
    NotFoundErrorTarefa:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          description: Mensagem explicativa do erro
          example: Nenhuma tarefa encontrada.
      additionalProperties: false
...
    ListarTarefa200:
      type: object
      properties:
        totalCount:
          type: integer
          description: Número total de tarefas que correspondem aos filtros aplicados.
        totalPages:
          type: integer
          description: "Número total de páginas, baseado no 'limit' fornecido."
        currentPage:
          type: integer
          description: Página atual (calculada com o offset e limit).
        data:
          maxItems: 50
          type: array
          items:
            $ref: "#/components/schemas/Tarefa"
      additionalProperties: false
    ListarUmaTarefa200:
      type: object
      properties:
        data:
          maxItems: 50
          type: array
          items:
            $ref: "#/components/schemas/Tarefa"
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT